Check temperature sensor file(s) exist on start up.
With the change to the new-style lm-sensors package providing the temperature information we now have the risk of a file not existing at start-up if the sensors library is not loaded or enumeration of devices is different. To avoid the risk of unplanned reboots we check first and disable any found to be missing.