Full asynchronous execution of test binaries
This rather large patch fixes a bug where multiple test/repair scripts with different execution times could result in their error codes being mixed up.

It changes from checking the exit status after starting a program to doing it first, so it gathers any children already finished and marks them off in the PID list. Then it looks in the list to see if you have already run the program and return its status if available, otherwise it runs it. If already running then it return the EDONTKNOW value to signify a running test.