Bugfix against watchdog configuration file corruption
- Apply the bugfix/patch against watchdog configuration file
  corruption during runtime by Greg Vishnepolsky greg@armis.com

- Modify from Greg's patch to include the permission mode and to
  declare the new fd variables immediately after the opening brace.

- Greg provided the follow description:

The following patch fixes a bug where writes/prints (out of a test
binary) that are intended for stdout/stderr may instead be written
to other files (like the watchdog configuration files).

The bug lies in the call to "freopen" on "stdout" when the actual
FD (1) for stdout had been previously closed by the daemon. Instead
this FD could have been re-used (for instance, for opening watchdog
configuration files). Thus any prints out of the daemon/test binaries
will be erroneously written into it.